wipwup

+++ lib/Command.dart
@@ -1,5 +1,5 @@
 class Command {
-  double outDataLength;
+  int outDataLength;
   List<int> inData;
 
   Command(this.inData, {this.outDataLength = 1024});
